# MM-27456: Use reflect-free serialization for hot structs