@mattermod 2020-08-02
The file `Makefile` is in the blocklist and should not be modified from external contributors, please if you are part of the Mattermost Org submit this PR in the upstream.
 /cc @mattermost/core-security @mattermost/core-build-engineers